// nims-web-app/src/app/core/models/sys-property-details.model.ts
import { SysPropertyBean } from './sys-property.model';
import { Nation } from './nation.model';
// Generated by https://quicktype.io

export class SysPropertyDetailBean {
    constructor(_sysProperty: any, _nation: any) {
        this.propertyId = _sysProperty.propertyId;
        this.propertyCode = _sysProperty.code;
        this.propertyName = _sysProperty.name;
        this.tableName = _sysProperty.tableName;
        this.columnName = _sysProperty.columnName;
        this.nationId = _nation.nationId;
        this.nationCode = _nation.code;
        this.nationName = _nation.name;
    }
    propertyCode?: string;
    propertyName?: string;
    nationCode?: string;
    nationName?: string;
    propertyDetailId?: number;
    isHide?:           boolean;
    isTranslation?:    boolean;
    align?:            string;
    dateFormat?:       string;
    numberFormat?:     string;
    moneyFormat?:      string;
    css?:              string;
    js?:               string;
    propertyId?:       number;
    nationId?:         number;
    isRequire?:        boolean;
    isUrl?:            boolean;
    isNumber?:         boolean;
    isEmail?:          boolean;
    isIp?:             boolean;
    minLength?:        number;
    maxLength?:        number;
    numberMin?:        number;
    numberMax?:        number;
    tableName?:        string;
    columnName?:       string;
  }